requirements.txt
torch>=2.0.0
transformers>=4.36.0
datasets>=2.14.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.24.0
trl>=0.7.0
unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git
azureml-core>=1.54.0
azureml-train-core>=1.54.0
azureml-train-automl-client>=1.54.0
azureml-pipeline>=1.54.0
azureml-defaults>=1.54.0
mlflow>=2.8.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
wandb>=0.16.0
tensorboard>=2.14.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
pyyaml>=6.0
jsonlines>=3.1.0